fix(audit): register audit sources in make build + split to stay under line limit
Fixes the S2 CI failures:
- Production make build: add audit_action.c and the new guardrails_action_audit.c
  to the Makefile source lists (guardrails_orchestrator referenced audit_args_hash
  but the make build's explicit lists didn't compile it -> undefined reference in
  aimee-server / unit-test-agent). CMake DATA_SRCS updated to match.
- line-check gate (guardrails_orchestrator.c was exactly at the 2000-line limit):
  move ALL the S2 audit wiring out into a new guardrails_action_audit.c. The
  verdict logic becomes pre_tool_check_inner (declared in guardrails.h); the thin
  pre_tool_check wrapper + emit live in the new file. guardrails_orchestrator.c is
  a pure rename -> back to 2000 lines.
- reason_code now derives from the block site's EXISTING audit_log("<key>",...)
  via a thread-local last-event captured in log.c (audit_last_event) — zero added
  lines at the block sites. computer-use approval_required detected via the
  msg_buf prefix (classification only; prose never persisted).
