+++ src/headers/guardrails.h
@@ -138,6 +138,13 @@ classification_t classify_path(const char *file_path);
 int pre_tool_check(const char *tool_name, const char *input_json, session_state_t *state,
                    const char *guardrail_mode, const char *cwd, char *msg_buf, size_t msg_len);
 
+/* The guardrail verdict logic. pre_tool_check is a thin wrapper (in
+ * guardrails_action_audit.c) that calls this and emits the per-action audit
+ * row. Same return contract as pre_tool_check. */
+int pre_tool_check_inner(const char *tool_name, const char *input_json, session_state_t *state,
+                         const char *guardrail_mode, const char *cwd, char *msg_buf,
+                         size_t msg_len);
+
 /* Normalize provider/internal tool names into guardrail-facing categories.
  * Provider-native sub-agent spawns (Task/Agent/spawn_agent/RemoteTrigger) map to
  * "Subagent" — callers test for that to detect a sub-agent call. */
